+++ content/__tests__/bootstrap.test.ts
@@ -2,15 +2,15 @@
 
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { setupLeetcodeAutoReset } from '@/content/auto-reset';
+import { sendMessage } from '@/infrastructure/browser/messages';
 import { getStoredTranslations } from '@/infrastructure/storage/translations';
 import { translations } from '@/shared/i18n';
-import { sendMessage } from '@/shared/messages';
 import { createDeferred } from '@/test/utils/deferred';
 import { bootstrapContent } from '../bootstrap';
 
 vi.mock('@/content/auto-reset', () => ({ setupLeetcodeAutoReset: vi.fn() }));
 vi.mock('@/infrastructure/storage/translations', () => ({ getStoredTranslations: vi.fn() }));
-vi.mock('@/shared/messages', () => ({ sendMessage: vi.fn() }));
+vi.mock('@/infrastructure/browser/messages', () => ({ sendMessage: vi.fn() }));
 
 let notifyMutation: () => void;
 const observe = vi.fn();
