+++ entrypoints/popup/components/__tests__/StreakCounter.test.tsx
@@ -4,13 +4,13 @@ import { Rating } from 'ts-fsrs';
 import { describe, expect, it, vi } from 'vitest';
 import type { DailyStats } from '@/domain/stats';
 import { statsQueryKeys } from '@/entrypoints/popup/queries/stats';
-import { sendMessage } from '@/shared/messages';
+import { sendMessage } from '@/infrastructure/browser/messages';
 import { createDeferred } from '@/test/utils/deferred';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { StreakCounter } from '../StreakCounter';
 
-vi.mock('@/shared/messages', () => ({ sendMessage: vi.fn() }));
+vi.mock('@/infrastructure/browser/messages', () => ({ sendMessage: vi.fn() }));
 
 const stats = (streak: number): DailyStats => ({
   date: '2024-03-15',
