+++ entrypoints/popup/views/stats/__tests__/ReviewHistoryChart.test.tsx
@@ -7,7 +7,7 @@ import { Rating } from 'ts-fsrs';
 import { describe, expect, it, vi } from 'vitest';
 import type { DailyStats } from '@/domain/stats';
 import { statsQueryKeys } from '@/entrypoints/popup/queries/stats';
-import { sendMessage } from '@/shared/messages';
+import { sendMessage } from '@/infrastructure/browser/messages';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { ReviewHistoryChart } from '../ReviewHistoryChart';
@@ -21,7 +21,7 @@ vi.mock('react-chartjs-2', () => ({
   ),
 }));
 
-vi.mock('@/shared/messages', () => ({ sendMessage: vi.fn() }));
+vi.mock('@/infrastructure/browser/messages', () => ({ sendMessage: vi.fn() }));
 
 describe('Bar Chart (Last 30 Days Review History)', () => {
   const messages = createMessageMock(vi.mocked(sendMessage));
